Restructure to put operations in own files + add bulkImport dir
